// apps/silo/src/apps/slack/types/workobjects.ts
export interface IIssueDetailService {
  getWorkItemDetails(
    planeClient: Client,
    ctx: TWorkItemDetailsCtx & { workspaceSlug: string }
  ): Promise<TWorkObjectIssueDetails>;
}

export interface IIssueWorkObjectViewService {
  getWorkItemView(
    workspaceSlug: string,
    issueDetails: TWorkObjectIssueDetails,
    config?: TWorkObjectViewConfig
  ): TWorkObjectView;
}

/* -------------------- TYPES -------------------------- */
export type TWorkObjectType = "MINIMAL" | "DETAILED"; // WHY NOT USE AN ENUM OVER HERE?

export enum EWorkObjectEntityType {
  FILE = "slack#/entities/file",
  TASK = "slack#/entities/task",
  ITEM = "slack#/entities/item",
  CONTENT_ITEM = "slack#/entities/content_item",
  INCIDENT = "slack#/entities/incident",
}

export type TWorkItemDetailsCtx =
  | { strategy: "sequence"; projectIdentifier: string; issueSequence: number }
  | { strategy: "id"; projectId: string; issueId: string };
// ...
export type TWorkObjectEntityPayload = {
  attributes: TWorkObjectEntityPayloadAttributes;
  fields: Record<string, TWorkObjectField & TWorkObjectFieldValue>;
  custom_fields: Array<TWorkObjectField & TWorkObjectCustomFieldValue>;
  display_order: string[];
  actions?: {
    primary_actions: TWorkObjectAction[];
    overflow_actions: TWorkObjectAction[];
  };
};

export type TWorkObjectView = {
  app_unfurl_url?: string;
  url: string;
  external_ref: {
    id: string;
    type: string;
  };
  entity_type: EWorkObjectEntityType;
  entity_payload: TWorkObjectEntityPayload;
};

export type TWorkObjectViewConfig = {
  appUnfurlUrl?: string;
};

/**
 * Additional details that can be fetched for an issue work object
 * Each property here should have a corresponding fetcher in IssueDetailService
 */
export type TWorkObjectAdditionalIssueDetails = {
  propertyDetails?: {
    properties: ExIssueProperty[];
    propertyOptions: Map<string, ExIssuePropertyOption[]>;
    propertyValues: Map<string, { property_id: string; values: string[] }[]>;
  };
  availableStates?: ExState[];
  projectMembers?: PlaneUser[];

  // Issue Properties, that we need to show
  parent?: ExIssue;
};

export type TWorkObjectIssueDetails = IssueWithExpanded<
  ["state", "project", "assignees", "labels", "type", "created_by", "updated_by"]
> & {
  additionalDetails?: TWorkObjectAdditionalIssueDetails;
};

export type TWorkObjectContext = {
  planeCtx: {
    workspaceId: string;
    workspaceSlug: string;
  };
  planeClient: Client;
  userMap: {
    planeToSlack: Map<string, string>;
    slackToPlane: Map<string, string>;
  };
};